First 100 Fibonacci numbers

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -10,20 +10,20 @@
 int main(void)
 {
 	int i = 0;
-	int j = 1;
-	int k = 2;
+	long j = 1;
+	long k = 2;
 
 	while (i < 50)
 	{
 	if (i == 0)
-		printf("%d", j);
+		printf("%ld", j);
 	else if (i == 1)
-		printf(", %d", k);
+		printf(", %ld", k);
 	else
 	{
 		k += j;
 		j = k - j;
-		printf(", %d", k);
+		printf(", %ld", k);
 	}
 	++i;
